feat: переносит функцию создания случайных чисел в отдельный модуль
