+++ js/utilities.js
@@ -1,0 +1,9 @@
+/*фунция получения псевдослучайного числа*/
+const getRandomNumber = (a, b) => {
+  const min = Math.min(a, b);
+  const max = Math.max(a, b);
+  const result = Math.floor(Math.random() * (max - min + 1)) + min;
+  return result;
+};
+
+export {getRandomNumber};
